+++ src/kodo/rlnc/full_rlnc_codes.hpp
@@ -14,3 +14,5 @@
 #include "shallow_backward_full_rlnc_decoder.hpp"
 #include "shallow_delayed_full_rlnc_decoder.hpp"
 #include "shallow_sparse_full_rlnc_encoder.hpp"
+
+#include "sparse_full_rlnc_encoder.hpp"