+++ test/src/test_sparse_full_rlnc_encoder.cpp
@@ -1,0 +1,8 @@
+// Copyright Steinwurf ApS 2011.
+// Distributed under the "STEINWURF RESEARCH LICENSE 1.0".
+// See accompanying file LICENSE.rst or
+// http://www.steinwurf.com/licensing
+
+/// @file test_sparse_full_rlnc_encoder.cpp The unit tests for the
+///       sparse_full_rlnc_encoder stack is defined in the
+///       test_full_rlnc_codes.cpp